Additions to the document

Functions to add content to de document:

add_cur_head() adds the heading of the current section, with the specified level, and customized text. add_subhead is passed to a 'headx' block, to add it to the doc, with a new heading level, and possibly accompanying the section header..

add_custom() adds any custom text.

add_pause() adds '. . .', add_br() adds '<br>', and add_lb() adds '\n\n'. get_end() is a wrapper to choose between them.

Usage

add_cur_head(n = NULL, x = "")

add_subhead(n = NULL, sep = " - ", x = "", sub = TRUE)

add_custom(x, end = "pbr")

add_pause(br = TRUE, trailing = TRUE)

add_br(trailing = TRUE)

add_lb()

get_end(end, ...)

Arguments

n

The heading level to use. Set to NA to ommit. Set to NULL (the default) to use the value of head_lv.

x

The custom text to add.

sep

If sub = TRUE, the separator to use between the section headers.

sub

Should the header be added to the current section header, or be a standalone header?

end

The type of end to add. See get_end().

br

Should a '<br>' be added after the pause?

trailing

Should an empty line be added before?

...

Arguments to pass to the add_* functions.