Edit Blocks

These functions are assigned to blocks inside plan_sec(b* = fun(...)). See planning. e and div edit existing blocks, the rest adds new markdown content.

e is used on single element blocks ('text expressions'). It splits the expression into clauses, and allows the user to select and alter each.

div is used on multi-element blocks. Currently, it returns the block as is.

add_cur_head adds the heading of the current section, with the specified level.

add_pause adds '. . .' and add_br adds '<br>'.

add_custom adds any custom text.

Usage

e(
  cs = NULL,
  end = "pbr",
  breaks = NULL,
  adds = NULL,
  subs = NULL,
  pat = "(?<=[\\.:;!?])(?<!\\{\\.)",
  sep = add_pause()
)

div(cs = TRUE, end = "p", breaks = NULL, sep = add_pause())

add_cur_head(n = NULL)

add_custom(x, end = "pbr")

add_pause(br = TRUE)

add_br()

Arguments

cs

Integer vector of clauses (for e) or lines (for div) to keep. Defaults to all.

end

Character specifying what to add at the end of the block: 'pbr' for add_pause() (the default); 'p' for add_pause(FALSE), and 'br' for add_br().

breaks

Indexes to insert sep.

adds, subs

Character vector specifying additions or substitutions (respectively) on the final punctuation of each clause in cs. Defaults to none.

pat

The pattern to break expressions into clauses. Defaults to "(?<=[\\.:;!?])".

sep

What to add after the indexes specified by breaks.

n

The heading level to use.

x

The custom text to add.

br

Should a '<br>' be added after the pause?